requirements.txt
requests==2.31.0
PyJWT==2.8.0
PyYAML==6.0.1
rich==13.7.0
